auth layout cleanup (#2597)
