+++ web/pages/dashboard.tsx
@@ -1,7 +1,7 @@
 import { User } from "@supabase/auth-helpers-react";
 import { ReactElement } from "react";
 
-import AuthLayout from "../components/layout/authLayout";
+import AuthLayout from "../components/layout/auth/authLayout";
 import DashboardPage from "../components/templates/dashboard/dashboardPage";
 import { withAuthSSR } from "../lib/api/handlerWrappers";
 import {
