+++ web/pages/datasets/index.tsx
@@ -1,4 +1,4 @@
-import AuthLayout from "../../components/layout/authLayout";
+import AuthLayout from "../../components/layout/auth/authLayout";
 import { withAuthSSR } from "../../lib/api/handlerWrappers";
 import { User } from "@supabase/auth-helpers-react";
 import { SortDirection } from "../../services/lib/sorts/requests/sorts";
