+++ web/pages/evals/index.tsx
@@ -1,6 +1,6 @@
 import { User } from "@supabase/auth-helpers-react";
 import { ReactElement } from "react";
-import AuthLayout from "../../components/layout/authLayout";
+import AuthLayout from "../../components/layout/auth/authLayout";
 import EvalsPage from "../../components/templates/evals/EvalsPage";
 import { withAuthSSR } from "../../lib/api/handlerWrappers";
 import { SortDirection } from "../../services/lib/sorts/requests/sorts";
