+++ web/pages/prompts/[id]/index.tsx
@@ -2,7 +2,7 @@ import { User } from "@supabase/auth-helpers-react";
 import { GetServerSidePropsContext } from "next";
 
 import { ReactElement } from "react";
-import AuthLayout from "../../../components/layout/authLayout";
+import AuthLayout from "../../../components/layout/auth/authLayout";
 import PromptIdPage from "../../../components/templates/prompts/id/promptIdPage";
 import { SupabaseServerWrapper } from "../../../lib/wrappers/supabase";
