+++ web/pages/properties.tsx
@@ -1,5 +1,5 @@
 import { User } from "@supabase/auth-helpers-react";
-import AuthLayout from "../components/layout/authLayout";
+import AuthLayout from "../components/layout/auth/authLayout";
 import PropertiesPage from "../components/templates/properties/propertiesPage";
 import { withAuthSSR } from "../lib/api/handlerWrappers";
 import { ReactElement } from "react";
