+++ web/pages/test.tsx
@@ -2,7 +2,7 @@ import { GetServerSidePropsContext, InferGetServerSidePropsType } from "next";
 
 import { ReactElement } from "react";
 import { NextPageWithLayout } from "./_app";
-import AuthLayout from "../components/layout/authLayout";
+import AuthLayout from "../components/layout/auth/authLayout";
 import { SupabaseServerWrapper } from "../lib/wrappers/supabase";
 import getStripe from "../utlis/getStripe";
 import { useOrg } from "../components/layout/organizationContext";
